feat: 一括実行するジョブと最新を取得するジョブに分ける
